@@ -1,48 +1,52 @@
 {
   "name": "demschooltools",
   "dependencies": {
-    "@date-io/moment": "^3.0.0",
-    "@emotion/react": "^11.11.4",
-    "@emotion/styled": "^11.11.5",
-    "@mui/icons-material": "^5.15.18",
-    "@mui/material": "^5.15.18",
-    "@mui/styled-engine-sc": "^6.0.0-alpha.18",
-    "@mui/x-date-pickers": "^7.5.1",
-    "axios": "^1.7.2",
-    "bootstrap": "3.3.7",
-    "dayjs": "^1.11.11",
-    "handlebars": "4.0.10",
-    "jquery": "3.2.1",
-    "jquery-form": "4.2.1",
-    "jquery-ui": "1.12.1",
-    "localforage": "^1.7.3",
+    "@date-io/moment": "^3.2.0",
+    "@emotion/react": "^11.14.0",
+    "@emotion/styled": "^11.14.1",
+    "@mui/icons-material": "^7.3.2",
+    "@mui/material": "^7.3.2",
+    "@mui/styled-engine-sc": "^7.3.2",
+    "@mui/x-date-pickers": "^8.11.3",
+    "axios": "^1.12.2",
+    "bootstrap": "5.3.8",
+    "dayjs": "^1.11.18",
+    "handlebars": "4.7.8",
+    "jquery": "3.7.1",
+    "jquery-form": "4.3.0",
+    "jquery-ui": "1.14.1",
+    "localforage": "^1.10.0",
     "moment": "^2.30.1",
-    "react": "^18.3.1",
-    "react-icons": "^5.2.1",
-    "react-router-dom": "^6.23.1",
+    "react": "^19.1.1",
+    "react-dom": "^19.1.1",
+    "react-icons": "^5.5.0",
+    "react-router-dom": "^7.9.1",
     "shuffle-seed": "^1.1.6",
-    "spectrum-colorpicker": "1.8.0",
-    "uuid": "^9.0.1"
+    "spectrum-colorpicker": "1.8.1",
+    "uuid": "^13.0.0"
+  },
+  "devDependencies": {
+    "@babel/core": "^7.28.4",
+    "@babel/preset-env": "^7.28.3",
+    "@babel/preset-react": "^7.27.1",
+    "babel-loader": "^10.0.0",
+    "css-loader": "^7.1.2",
+    "esbuild-loader": "^4.3.0",
+    "eslint": "^8.57.1",
+    "eslint-config-standard": "^17.1.0",
+    "eslint-plugin-import": "^2.32.0",
+    "eslint-plugin-n": "^16.6.2",
+    "eslint-plugin-promise": "^6.6.0",
+    "eslint-webpack-plugin": "^4.2.0",
+    "html-webpack-plugin": "^5.6.4",
+    "sass": "^1.93.0",
+    "sass-loader": "^16.0.5",
+    "style-loader": "^4.0.0",
+    "webpack": "^5.101.3",
+    "webpack-cli": "^6.0.1"
   },
   "scripts": {
     "compile": "webpack --mode production",
     "watch": "webpack --watch --mode development"
-  },
-  "devDependencies": {
-    "@babel/core": "^7.24.6",
-    "@babel/preset-env": "^7.24.6",
-    "@babel/preset-react": "^7.24.6",
-    "babel-loader": "^9.1.3",
-    "css-loader": "^7.1.2",
-    "eslint": "^8.39.0",
-    "eslint-config-standard": "^17.0.0",
-    "eslint-webpack-plugin": "^4.0.1",
-    "html-webpack-plugin": "^5.5.1",
-    "react-dom": "^18.3.1",
-    "sass": "^1.77.2",
-    "sass-loader": "^14.2.1",
-    "style-loader": "^4.0.0",
-    "webpack": "^5.91.0",
-    "webpack-cli": "^5.1.4"
   }
 }
