@@ -8,34 +8,38 @@ const person_template_str =
 
 const person_template = Handlebars.compile(person_template_str);
 
-const Person = function(id, name, onClickPerson) {
+const Person = function (id, name, onClickPerson) {
     this.name = name;
     this.id = id;
     const self = this;
 
-    this.render = function() {
+    this.render = function () {
         return person_template({ name });
     };
 
     // called by PeopleChooser after self.el has been
     // initialized.
-    this.activateClick = function() {
+    this.activateClick = function () {
         if (onClickPerson) {
-            self.el.click(
-                function() {
-                    onClickPerson(self);
-                }
-            );
+            self.el.click(function () {
+                onClickPerson(self);
+            });
         }
     };
 };
 
-const PeopleChooser = function(el, on_add, on_remove, autocomplete_source, onClickPerson) {
+const PeopleChooser = function (
+    el,
+    on_add,
+    on_remove,
+    autocomplete_source,
+    onClickPerson,
+) {
     this.el = el;
     this.people = [];
     const self = this;
 
-    this.search_box = el.find(".person_search");
+    this.search_box = el.find('.person_search');
     this.search_box.autocomplete({
         source: autocomplete_source,
         delay: 0,
@@ -44,13 +48,13 @@ const PeopleChooser = function(el, on_add, on_remove, autocomplete_source, onCli
 
     self.one_person_mode = false;
 
-//  el.mouseleave(function() {
-//      if (self.people.length > 0) {
-//          self.search_box.fadeOut();
-//      } } );
-//  el.mouseenter(function() { self.search_box.fadeIn(); } );
-//
-    this.search_box.bind( "autocompleteselect", function(event, ui) {
+    //  el.mouseleave(function() {
+    //      if (self.people.length > 0) {
+    //          self.search_box.fadeOut();
+    //      } } );
+    //  el.mouseenter(function() { self.search_box.fadeIn(); } );
+    //
+    this.search_box.bind('autocompleteselect', function (event, ui) {
         const new_person = self.addPerson(ui.item.id, ui.item.label);
 
         if (on_add && new_person) {
@@ -61,7 +65,7 @@ const PeopleChooser = function(el, on_add, on_remove, autocomplete_source, onCli
         event.preventDefault(); // keep jquery from inserting name into textbox
     });
 
-    this.addPerson = function(id, name) {
+    this.addPerson = function (id, name) {
         // Don't add people who have already been added.
         for (const i in self.people) {
             if (id == self.people[i].id) {
@@ -71,27 +75,32 @@ const PeopleChooser = function(el, on_add, on_remove, autocomplete_source, onCli
 
         if (self.one_person_mode) {
             self.search_box.hide();
-            self.el.find(".glyphicon").hide();
+            self.el.find('.glyphicon').hide();
             utils.selectNextInput(self.search_box);
         }
 
         const p = new Person(id, name, onClickPerson);
         self.people.push(p);
-        p.el = self.el.find(".people").append(p.render()).children(":last-child");
+        p.el = self.el
+            .find('.people')
+            .append(p.render())
+            .children(':last-child');
         p.activateClick();
 
-        p.el.find("img").click(function() { self.removePerson(p); });
+        p.el.find('img').click(function () {
+            self.removePerson(p);
+        });
 
         return p;
     };
 
-    this.clear = function() {
+    this.clear = function () {
         while (self.people.length > 0) {
             self.removePerson(self.people[0]);
         }
     };
 
-    this.removePerson = function(person) {
+    this.removePerson = function (person) {
         $(person.el).remove();
 
         for (const i in self.people) {
@@ -102,7 +111,7 @@ const PeopleChooser = function(el, on_add, on_remove, autocomplete_source, onCli
 
         if (self.one_person_mode) {
             self.search_box.show();
-            self.el.find(".glyphicon").show();
+            self.el.find('.glyphicon').show();
         }
 
         if (on_remove) {
@@ -110,13 +119,13 @@ const PeopleChooser = function(el, on_add, on_remove, autocomplete_source, onCli
         }
     };
 
-    this.loadPeople = function(people) {
+    this.loadPeople = function (people) {
         for (const i in people) {
             self.addPerson(people[i].id, people[i].name);
         }
     };
 
-    this.setOnePersonMode = function(one_person_mode) {
+    this.setOnePersonMode = function (one_person_mode) {
         self.one_person_mode = one_person_mode;
         return self;
     };
@@ -125,4 +134,4 @@ const PeopleChooser = function(el, on_add, on_remove, autocomplete_source, onCli
 module.exports = {
     PeopleChooser,
     Person,
-}
+};
