@@ -2,7 +2,7 @@ import React from 'react';
 import { Button as MuiButton } from '@mui/material';
 
 const Button = (props) => {
-  const { variant = 'outlined', ...restOfProps} = props;
+  const { variant = 'outlined', ...restOfProps } = props;
   return (
     <MuiButton variant={variant} {...restOfProps} />
   );
