@@ -3,7 +3,7 @@ const constants = require("./appconstants");
 const base = require("./storebase");
 const actionCreator = require("./studentactioncreator");
 
-let students, today;
+let students;
 const studentDetails = {};
 
 const exports = Object.create(base);
@@ -15,10 +15,6 @@ exports.getStudents = function (force) {
     actionCreator.loadStudents();
     return [];
   }
-};
-
-exports.getToday = function () {
-  return today;
 };
 
 exports.getStudent = function (id, force) {
@@ -34,7 +30,6 @@ dispatcher.register(function (action) {
   switch (action.type) {
     case constants.studentEvents.LOADED:
       students = action.data.students;
-      today = action.data.today;
       exports.emitChange();
       break;
     case constants.studentEvents.STUDENT_SWIPED:
