@@ -1,3 +1,5 @@
+const $ = require("jquery");
+
 function getCookie(name) {
   const cookie = {};
   document.cookie.split(";").forEach(function (el) {
@@ -11,10 +13,17 @@ const CSRF_HEADER = {
   "X-CSRFToken": getCookie("csrftoken"),
 };
 
+const modifyUrl = (url) => {
+  if (url.indexOf("/") === 0) {
+    return `/custodia-api${url}`;
+  }
+  return `/custodia-api/${url}`;
+};
+
 const exports = {
   post: function (url, data) {
     return $.ajax({
-      url,
+      url: modifyUrl(url),
       method: "POST",
       contentType: "application/json",
       dataType: "json",
@@ -24,7 +33,7 @@ const exports = {
   },
   put: function (url, data) {
     return $.ajax({
-      url,
+      url: modifyUrl(url),
       method: "PUT",
       contentType: "application/json",
       dataType: "json",
@@ -32,12 +41,15 @@ const exports = {
       headers: CSRF_HEADER,
     });
   },
-  get: function (url) {
-    return $.ajax(url);
+  get: function (inObj) {
+    if (typeof inObj === "string") {
+      return $.ajax(modifyUrl(inObj));
+    }
+    return $.ajax(Object.assign({}, inObj, { url: modifyUrl(inObj.url) }));
   },
   delete: function (url) {
     return $.ajax({
-      url,
+      url: modifyUrl(url),
       method: "DELETE",
       headers: CSRF_HEADER,
     });
