@@ -1,5 +1,6 @@
 const React = require("react");
 const Heatmapmonth = require("./heatmapmonth.jsx");
+const dayjs = require("dayjs");
 
 Array.prototype.concatAll = function () {
   const results = [];
@@ -15,37 +16,50 @@ const groupingFunc = function (data) {
   return data.day.split("-")[0] + "-" + data.day.split("-")[1] + "-" + "01";
 };
 
-module.exports = class extends React.Component {
-  static displayName = "Heatmap";
+module.exports = class Heatmap extends React.Component {
+  constructor(props) {
+    super(props);
+    this.state = {
+      mounted: false,
+    };
+
+    // Defer rendering the heatmaps so that the initial page load can take
+    // place without waiting for them.
+    setTimeout(() => {
+      this.setState({ mounted: true });
+    }, 0);
+  }
 
   loadHeatmaps = () => {
     const groupedDays = this.props.days.groupBy(groupingFunc);
     const minutes = this.props.requiredMinutes;
     const sortedDates = Object.keys(groupedDays).sort();
 
-    let i;
-    let j;
-    let temparray;
-    const chunk = 4;
+    let i = 0;
     const maps = [];
-    for (i = 0, j = sortedDates.length; i < j; i += chunk) {
-      temparray = sortedDates.slice(i, i + chunk);
-      const dates = temparray
+    while (i < sortedDates.length) {
+      // Find up to four months worth of dates, which we will send
+      // to Heatmapmonth in one batch.
+      let j = i + 1;
+      const lastDate = dayjs(sortedDates[i]).add(4, "months");
+      while (j < sortedDates.length && dayjs(sortedDates[j]).isBefore(lastDate)) {
+        j++;
+      }
+
+      const dates = sortedDates
+        .slice(i, j)
         .map(function (d) {
           return groupedDays[d];
         })
         .concatAll();
-      const m = this.makeHeatmapRow(dates, minutes, i);
-      maps.push(m);
+      i = j;
+
+      maps.push(<Heatmapmonth key={i} index={i} requiredMinutes={minutes} days={dates} />);
     }
     return maps;
   };
 
-  makeHeatmapRow = (dates, minutes, key) => {
-    return <Heatmapmonth key={key} index={key} requiredMinutes={minutes} days={dates} />;
-  };
-
   render() {
-    return <div className="row">{this.loadHeatmaps()}</div>;
+    return this.state.mounted && <div className="row">{this.loadHeatmaps()}</div>;
   }
 };
