@@ -1,27 +1,15 @@
-"""
-Django settings for demschooltools project.
-
-Generated by 'django-admin startproject' using Django 3.2.21.
-
-For more information on this file, see
-https://docs.djangoproject.com/en/3.2/topics/settings/
-
-For the full list of settings and their values, see
-https://docs.djangoproject.com/en/3.2/ref/settings/
-"""
-
 import os
 
-from settings import *  # noqa: F403
+from demschooltools.settings import *  # noqa: F403
 
 DATABASES = {
     "default": {
         "ENGINE": "django.db.backends.postgresql",
-        "OPTIONS": {"options": "-c search_path=public,overseer"},
         "NAME": "school_crm",
         "PORT": "5433",
         "HOST": "localhost",
         "USER": "evan",
         "PASSWORD": os.environ["DB_PASSWORD"],
+        "CONN_MAX_AGE": 60,
     },
 }
