@@ -11,6 +11,6 @@ import os
 
 from django.core.wsgi import get_wsgi_application
 
-os.environ.setdefault("DJANGO_SETTINGS_MODULE", "demschooltools.settings")
+assert "DJANGO_SETTINGS_MODULE" in os.environ
 
 application = get_wsgi_application()
