@@ -46,7 +46,7 @@ public class Account extends Model {
   private Date dateLastMonthlyCredit;
 
   public static void createPersonalAccounts(Organization org) {
-    for (Person person : allPeople(org)) {
+    for (Person person : Person.all(org)) {
       if (!person.hasAccount(AccountType.PersonalChecking)) {
         create(AccountType.PersonalChecking, "", person, org);
       }
@@ -71,13 +71,13 @@ public class Account extends Model {
     return "<no name>";
   }
 
-  public String getTitle(OrgConfig orgConfig) {
-    String typeName = type == AccountType.PersonalChecking ? getTypeName(orgConfig) + " " : "";
+  public String getTitle() {
+    String typeName = type == AccountType.PersonalChecking ? getTypeName() + " " : "";
     return getName() + "'s " + typeName + "Account";
   }
 
-  public String getTypeName(OrgConfig orgConfig) {
-    return type.toString(orgConfig);
+  public String getTypeName() {
+    return type.toString(organization);
   }
 
   public boolean hasTransactions() {
