@@ -4,10 +4,7 @@ import com.feth.play.module.pa.user.AuthUser;
 import com.feth.play.module.pa.user.AuthUserIdentity;
 import io.ebean.*;
 import io.ebean.ExpressionList;
-import java.util.Arrays;
-import java.util.HashSet;
 import java.util.List;
-import java.util.Set;
 import javax.persistence.*;
 import lombok.Getter;
 import lombok.Setter;
@@ -18,6 +15,8 @@ import lombok.Setter;
 @Table(name = "users")
 public class User extends Model {
   public static final String DUMMY_USERNAME = "__DUMMY_USERNAME__";
+  public static final String CHECKIN_USERNAME = "Check-in app user";
+
   private static final long serialVersionUID = 1L;
 
   @ManyToOne() private Organization organization;
@@ -55,7 +54,9 @@ public class User extends Model {
 
     result.save();
 
-    return result;
+    // User.linkedAccounts is only non-null if this object has been loaded from
+    // the database, so get the full thing from the DB to return.
+    return User.findById(result.id, org);
   }
 
   public boolean hasRole(String role) {
@@ -84,56 +85,6 @@ public class User extends Model {
       return User.findByEmail(identity.getId());
     }
     return getAuthUserFind(identity).findOne();
-  }
-
-  public void merge(final User otherUser) {
-    for (final LinkedAccount acc : otherUser.linkedAccounts) {
-      this.linkedAccounts.add(LinkedAccount.create(acc));
-    }
-    // do all other merging stuff here - like resources, etc.
-
-    // deactivate the merged user that got added to this one
-    otherUser.active = false;
-    DB.save(Arrays.asList(otherUser, this));
-  }
-
-  // public static User create(final AuthUser authUser) {
-  //	final User user = new User();
-  //	user.active = true;
-  //	user.linkedAccounts = Collections.singletonList(LinkedAccount
-  //			.create(authUser));
-  //
-  //	if (authUser instanceof EmailIdentity) {
-  //		final EmailIdentity identity = (EmailIdentity) authUser;
-  //		// Remember, even when getting them from FB & Co., emails should be
-  //		// verified within the application as a security breach there might
-  //		// break your security as well!
-  //		user.email = identity.getEmail();
-  //		user.emailValidated = false;
-  //	}
-  //
-  //	if (authUser instanceof NameIdentity) {
-  //		final NameIdentity identity = (NameIdentity) authUser;
-  //		final String name = identity.getName();
-  //		if (name != null) {
-  //			user.name = name;
-  //		}
-  //	}
-  //
-  //	user.save();
-  //	return user;
-  // }
-
-  public static void merge(final AuthUser oldUser, final AuthUser newUser) {
-    User.findByAuthUserIdentity(oldUser).merge(User.findByAuthUserIdentity(newUser));
-  }
-
-  public Set<String> getProviders() {
-    final Set<String> providerKeys = new HashSet<>(linkedAccounts.size());
-    for (final LinkedAccount acc : linkedAccounts) {
-      providerKeys.add(acc.getProviderKey());
-    }
-    return providerKeys;
   }
 
   public static void addLinkedAccount(final AuthUser oldUser, final AuthUser newUser) {
